+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Avalonia: Added support for Avalonia 11 (#2720, #2184, #2664, #2507, #2321, #2184, #2183, #2725) _Thanks @Fruchtzwerg94, @Xerxes004, and @bclehmann_
 * Colorbar: Added a `ResizeLayout()` helper method for adjusting plot layouts to accommodate large tick labels (#2703)
 * Scatter List: Improved support for data containing NaN values (#2707) _Thanks @oldteacup_
+* Population Plot: Improved support for populations with no data (#2727, #2726) _Thanks @marklam_
 
 ## ScottPlot 5.0.5-beta
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2023-05-17_